+++ manage.py
@@ -3,7 +3,7 @@ import sys
 
 from django.core.management import execute_from_command_line
 
-from wger.utils.main import (
+from wger.tasks import (
     setup_django_environment,
     get_user_config_path
 )
