+++ setup.cfg
@@ -1,4 +1,4 @@
 [pep8]
-exclude = urls.py,*migrations*,*bower_components*
+exclude = urls.py,tasks.py,*migrations*,*bower_components*
 max-line-length = 100
 ignore = W503
