+++ wger/settings_global.py
@@ -25,9 +25,12 @@ For a full list of options, visit:
 
 # Build paths inside the project like this: os.path.join(BASE_DIR, ...)
 import os
-BASE_DIR = os.path.dirname(os.path.dirname(__file__))
+BASE_DIR = os.path.abspath(os.path.dirname(__file__))
 SITE_ROOT = os.path.realpath(os.path.dirname(__file__))
 
+print(BASE_DIR)
+print(SITE_ROOT)
+print('===========================')
 
 #
 # Application definition
@@ -313,6 +316,8 @@ if sys.platform.startswith('win32'):
     BOWER_PATH = os.path.join(BASE_DIR, 'node_modules', '.bin', 'bower.cmd')
 else:
     BOWER_PATH = os.path.join(BASE_DIR, 'node_modules', '.bin', 'bower')
+
+print(BOWER_PATH)
 
 #
 # Django Rest Framework
